Implemented player movement, collisions, terrains, camera/camera following the player and isometric view.
Implemented the following:

Diagonal player movement for a 2D world, build on a isometric tilemap grid.
Movement input is handled and validate using the new input system.

Direction they're facing indicated by a small "nose" that gets rotated with the player.

There is up different terrain, that slows or speeds the player up, for them to interact with, as well as a barrier they cannot move through.

The camera uses the new built-in Cinemachine to follow the player, with default smoothing and framing settings for a stable isometric view.